+++ src/mako/config/exceptions/ConfigException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\config\exceptions;
+
+use RuntimeException;
+
+/**
+ * Config exception.
+ */
+class ConfigException extends RuntimeException
+{
+
+}