+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The major version bump is due to dropped support for PHP `7.3` and a several bre
 * Removed the deprecated `AccessControlAllowOrigin` middleware.
 * Moved the `mako\cli\output\formatter\FormatterException` exception to `mako\cli\output\formatter\exceptions\FormatterException`.
 * Moved the `mako\redis\RedisException` exception to `mako\redis\exceptions\RedisException`.
+* Moved the `mako\config\loaders\LoaderException` exception to `mako\config\loaders\exceptions\LoaderException`.
 
 
 --------------------------------------------------------